Separation des éléments de menu dans une autre fichier
